push client directive to leaf interactive components, keep content in server shell
