+++ apps/www/app/blog/[slug]/BlogPostClient.tsx
@@ -1,13 +1,8 @@
-'use client'
-
-import dynamic from 'next/dynamic'
-
-import useActiveAnchors from '@/hooks/useActiveAnchors'
+import BlogPostAnchorEffect from './BlogPostAnchorEffect'
+import BlogPostRenderer from '@/components/Blog/BlogPostRenderer'
 import authors from '@/lib/authors.json'
 import { isNotNullOrUndefined } from '@/lib/helpers'
 import type { Blog, BlogData, PostReturnType, ProcessedBlogData } from '@/types/post'
-
-const BlogPostRenderer = dynamic(() => import('@/components/Blog/BlogPostRenderer'))
 
 type BlogPostPageProps = {
   prevPost: PostReturnType | null
@@ -17,14 +12,7 @@ type BlogPostPageProps = {
   isDraftMode: boolean
 }
 
-// Note: convertRichTextToMarkdown is now imported from the shared utility
-
 export default function BlogPostClient(props: BlogPostPageProps) {
-  const isDraftMode = props.isDraftMode
-
-  // Enable scroll-to-anchor functionality and TOC highlighting
-  useActiveAnchors()
-
   const blogMetaData = props.blog
 
   const blogAuthors = (blogMetaData.author ?? '')
@@ -38,13 +26,16 @@ export default function BlogPostClient(props: BlogPostPageProps) {
     .filter(isNotNullOrUndefined)
 
   return (
-    <BlogPostRenderer
-      blog={props.blog as ProcessedBlogData}
-      blogMetaData={blogMetaData as ProcessedBlogData}
-      isDraftMode={isDraftMode}
-      prevPost={props.prevPost}
-      nextPost={props.nextPost}
-      authors={blogAuthors}
-    />
+    <>
+      <BlogPostAnchorEffect />
+      <BlogPostRenderer
+        blog={props.blog as ProcessedBlogData}
+        blogMetaData={blogMetaData as ProcessedBlogData}
+        isDraftMode={props.isDraftMode}
+        prevPost={props.prevPost}
+        nextPost={props.nextPost}
+        authors={blogAuthors}
+      />
+    </>
   )
 }
